style: infra 상위 패키지 pennyway 추가 -> 디렉토리 이동
