+++ pennyway-app-external-api/src/main/java/kr/co/pennyway/api/common/security/jwt/access/AccessTokenProvider.java
@@ -4,13 +4,13 @@ import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.JwtException;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
-import kr.co.infra.common.exception.JwtErrorCode;
-import kr.co.infra.common.exception.JwtErrorException;
-import kr.co.infra.common.jwt.JwtClaims;
-import kr.co.infra.common.jwt.JwtProvider;
-import kr.co.infra.common.util.JwtErrorCodeUtil;
 import kr.co.pennyway.api.common.annotation.AccessTokenStrategy;
 import kr.co.pennyway.common.util.DateUtil;
+import kr.co.pennyway.infra.common.exception.JwtErrorCode;
+import kr.co.pennyway.infra.common.exception.JwtErrorException;
+import kr.co.pennyway.infra.common.jwt.JwtClaims;
+import kr.co.pennyway.infra.common.jwt.JwtProvider;
+import kr.co.pennyway.infra.common.util.JwtErrorCodeUtil;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Primary;
