+++ pennyway-app-external-api/src/test/java/kr/co/pennyway/api/common/security/jwt/access/AccessTokenProviderTest.java
@@ -1,9 +1,9 @@
 package kr.co.pennyway.api.common.security.jwt.access;
 
-import kr.co.infra.common.exception.JwtErrorCode;
-import kr.co.infra.common.exception.JwtErrorException;
-import kr.co.infra.common.jwt.JwtClaims;
-import kr.co.infra.common.jwt.JwtProvider;
+import kr.co.pennyway.infra.common.exception.JwtErrorCode;
+import kr.co.pennyway.infra.common.exception.JwtErrorException;
+import kr.co.pennyway.infra.common.jwt.JwtClaims;
+import kr.co.pennyway.infra.common.jwt.JwtProvider;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
