+++ pennyway-infra/src/main/java/kr/co/pennyway/infra/client/aws/sms/SmsDto.java
@@ -1,4 +1,4 @@
-package kr.co.infra.client.aws.sms;
+package kr.co.pennyway.infra.client.aws.sms;
 
 import lombok.Builder;
