+++ pennyway-infra/src/main/java/kr/co/pennyway/infra/common/exception/JwtErrorCode.java
@@ -1,4 +1,4 @@
-package kr.co.infra.common.exception;
+package kr.co.pennyway.infra.common.exception;
 
 import kr.co.pennyway.common.exception.BaseErrorCode;
 import kr.co.pennyway.common.exception.CausedBy;
