+++ pennyway-infra/src/main/java/kr/co/pennyway/infra/common/jwt/JwtClaims.java
@@ -1,4 +1,4 @@
-package kr.co.infra.common.jwt;
+package kr.co.pennyway.infra.common.jwt;
 
 import java.util.Map;
